+++ src/coding_assistant/agents/tests/helpers.py
